Fix export after merging pull request #2
The views have be to non-temporary, otherwise they only persist for the current session. So the view 'xml_stopplaces' cannot be accessed in the second session after running the python script.

The second creation of the 'category' type was removed, because this drops it's related views/tables that are used further down.